+++ .streamlit/config.toml
@@ -3,6 +3,7 @@ developmentMode = false
 
 [server]
 port = 8501 # should be same as configured in deployment repo
+maxUploadSize = 1000
 
 [theme]
