+++ src/parse/quant.py
@@ -1,5 +1,5 @@
-from src.flashquant import parseFLASHQuantOutput
-from src.flashquant import connectTraceWithResult
+from src.parse.flashquant import parseFLASHQuantOutput
+from src.parse.flashquant import connectTraceWithResult
 
 
 def parseQuant(quant_tsv, trace_tsv, conflict_tsv=None):
