+++ src/render/util.py
@@ -1,0 +1,6 @@
+import pickle
+import hashlib
+
+def hash_complex(d):
+    serialized = pickle.dumps(d)
+    return hashlib.sha256(serialized).hexdigest()
